Updated pages to recognize when an account is logged in. Implemented sign out button in Navbar.jsx
