+++ Coding-Crazy/src/App.jsx
@@ -2,7 +2,7 @@ import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Navbar from "./components/Navbar";
 import HomePage from "./components/HomePage";
 import GamePage from "./components/GamePage";
-import StudyPage from "./components/StudyPage";
+import AccountPage from "./components/AccountPage";
 import LobbyPage from "./components/LobbyPage";
 import HostPage from "./components/HostPage";
 import LoginPage  from "./components/LoginPage";
@@ -17,7 +17,7 @@ function App() {
                 <Routes>
                     <Route path="/" element={<HomePage />} />
                     <Route path="/game" element={<GamePage />} />
-                    <Route path="/study" element={<StudyPage />} />
+                    <Route path="/account" element={<AccountPage />} />
                     <Route path="/lobby/:accessCode" element={<LobbyPage />} />
                     <Route path="/setupGame" element={<HostPage />} />
                     <Route path="/login" element={<LoginPage />} />
