+++ Coding-Crazy/src/components/Navbar.jsx
@@ -1,10 +1,23 @@
-import { AppBar, Toolbar, IconButton, Typography, Button, Drawer, List, ListItem, ListItemText, Box } from "@mui/material";
+import { AppBar, Toolbar, IconButton, Typography, Button, Drawer, List, ListItem, ListItemText, Box, useStepContext } from "@mui/material";
 import MenuIcon from "@mui/icons-material/Menu";
 import { Link } from "react-router-dom";
-import { useState } from "react";
+import { useEffect, useState } from "react";
 
 function Navbar() {
     const [drawerOpen, setDrawerOpen] = useState(false);
+    const [isLoggedIn, setIsLoggedIn] = useState(localStorage.getItem("username") != null);
+
+    useEffect(() => {
+        const checkLoginStatus = () => {
+            setIsLoggedIn(localStorage.getItem("username") != null);
+        };
+    
+        window.addEventListener("storage", checkLoginStatus);   // If storage changes, check if account was effected
+    
+        return () => {
+            window.removeEventListener("storage", checkLoginStatus);
+        };
+    }, []);
 
     const toggleDrawer = (open) => (event) => {
         if (event.type === "keydown" && (event.key === "Tab" || event.key === "Shift")) {
@@ -12,6 +25,12 @@ function Navbar() {
         }
         setDrawerOpen(open);
     };
+
+    const signOut = () => {
+        localStorage.clear();
+        setIsLoggedIn(false);
+        console.log(`Signed out: ${localStorage.getItem("username")}`);
+    }
 
     return (
         <>
@@ -40,13 +59,21 @@ function Navbar() {
                         <Box sx={{ display: { xs: "none", sm: "flex" }, gap: 2 }}>
                             <Button color="inherit" component={Link} to="/">🏠 Home</Button>
                             <Button color="inherit" component={Link} to="/game">🎮 Play Game</Button>
-                            <Button color="inherit" component={Link} to="/study">Study Sets</Button>
+                            <Button color="inherit" component={Link} to="/account">Account</Button>
                             <Button color="inherit" component={Link} to="/setupGame">Setup Game</Button>
                         </Box>
                         {/* Log In & Sign Up Buttons (Desktop) */}
                         <Box sx={{ display: { xs: "none", sm: "flex" }, gap: 1 }}>
-                            <Button color="inherit" variant="outlined" component={Link} to="/login">Log In</Button>
-                            <Button color="success" variant="contained" component={Link} to="/signup">Sign Up</Button>
+                            { !isLoggedIn ? (
+                                <Box sx={{ display: { xs: "none", sm: "flex" }, gap: 1 }}>
+                                    <Button color="inherit" variant="outlined" component={Link} to="/login">Log In</Button>
+                                    <Button color="success" variant="contained" component={Link} to="/signup">Sign Up</Button>
+                                </Box>
+                            ) : (
+                                <Box sx={{ display: { xs: "none", sm: "flex" }, gap: 1 }}>
+                                    <Button color="inherit" variant="outlined" onClick={() => signOut()} component={Link} to="/">Sign Out</Button>
+                                </Box>
+                            )}
                         </Box>
                     </Box>
                 </Toolbar>
