+++ Coding-Crazy/src/components/SignupPage.jsx
@@ -1,11 +1,13 @@
 import { Box, Button, Typography, Grid, Card, CardContent, Container } from "@mui/material";
 import { useEffect, useState } from "react";
+import { Link } from "react-router-dom";
 
 function SignupPage() {
     /* IMPLEMENT EMAIL CHECKING TO PREVENT ACCOUNT CREATION SPAM */
     const [username, setUsername] = useState("");
     const [password, setPassword] = useState("");
     const [success, setSuccess] = useState("");
+    const [isLoggedIn, setIsLoggedIn] = useState(localStorage.getItem("username") != null);
 
     const signupAccount = async () => {
         try {
@@ -39,6 +41,8 @@ function SignupPage() {
             console.log(`Great Success!\nUsername: ${data.username}\nPassword: ${data.password}\nID: ${data._id}`);
             localStorage.setItem("username", data.username);    // Set username is local storage
             console.log(`Local Storage: ${localStorage.getItem("username")}`);
+            setIsLoggedIn(true);
+            window.dispatchEvent(new Event("storage")); // Let event handler know that local storage has changed
             setSuccess("Account Successfully Created.");    // Update success message
         } catch (error) {
             console.error("Error creating account:", error);
@@ -48,38 +52,38 @@ function SignupPage() {
 
     return (
         <Box sx={{ bgcolor: "#0f172a", color: "white", minHeight: "100vh" }}>
-            {/* Input Section */}
-            <Box textAlign="center" py={5}>
-                {/* Potential pitfall with re-redner on each keystroke: https://react.dev/reference/react-dom/components/input#usage*/}
-                <div>
-                    <label>
-                        Username: <input value={username} placeholder="Input Username Here" onChange={event => setUsername(event.target.value)} />
-                    </label>
-                </div>
-                <div>
-                    <label>
-                        Password: <input value ={password} placeholder="Input Password Here" onChange={event => setPassword(event.target.value)} />
-                    </label>
-                </div>
-            </Box>
-
-            {/* Confirmation Section */}
-            <Box textAlign="center" py={5}>
-                <h3>Your Username: {username}</h3>
-                <h3>Your Paswword: {password}</h3>
-                <Button variant="contained" color="primary" sx={{ mx: 1 }} onClick={
-                    () => signupAccount()}>Submit Signup Info</Button>
-                <h2>{success || ""}</h2> {/* Display success message */}
-            </Box>
-
-            {/* Reference Formatting */}
-            <Box textAlign="center" py={5}>
-                <Typography variant="h3" color="primary">TEMP TITLE</Typography>
-                <Typography variant="subtitle1">Temp text.</Typography>
-                <Box mt={3}>
-                    <Button variant="contained" color="secondary" sx={{ mx: 1 }}>temp button</Button>
+            { isLoggedIn ? (
+                <Box textAlign="center" py={5}>
+                    <h1>You are signed in.</h1>
+                    <Button component={Link} to="/">Return to Home Page</Button>
                 </Box>
-            </Box>
+            ) : (
+                <Box>
+                    {/* Input Section */}
+                    <Box textAlign="center" py={5}>
+                        {/* Potential pitfall with re-redner on each keystroke: https://react.dev/reference/react-dom/components/input#usage*/}
+                        <div>
+                            <label>
+                                Username: <input value={username} placeholder="Input Username Here" onChange={event => setUsername(event.target.value)} />
+                            </label>
+                        </div>
+                        <div>
+                            <label>
+                                Password: <input value ={password} placeholder="Input Password Here" onChange={event => setPassword(event.target.value)} />
+                            </label>
+                        </div>
+                    </Box>
+
+                    {/* Confirmation Section */}
+                    <Box textAlign="center" py={5}>
+                        <h3>Your Username: {username}</h3>
+                        <h3>Your Paswword: {password}</h3>
+                        <Button variant="contained" color="primary" sx={{ mx: 1 }} onClick={
+                            () => signupAccount()}>Submit Signup Info</Button>
+                        <h2>{success || ""}</h2> {/* Display success message */}
+                    </Box>
+                </Box>
+            )}
 
             {/* Footer */}
             <Box sx={{ bgcolor: "#1e293b", mt: 5, py: 3, textAlign: "center" }}>
