+++ Coding-Crazy/src/data/export_to_mongo.json
@@ -1,6 +1,1 @@
-[
-  {
-    "username": "Hiiiiiiiiiiiiii",
-    "password": "Bye"
-  }
-]
+[{"roomCode":"AGUJSD","usernames":[],"players":{},"maxPlayers":2,"difficulty":5,"countdownStarted":false,"countdown":10,"APlusLoc":12}]
