+++ Coding-Crazy/src/data/update_to_mongo.json
@@ -1,1 +1,1 @@
-[{"roomCode":"ARSSUD","usernames":["Samuel","Jamie Rup"],"players":{"Samuel":{"id":"Samuel","loc":0,"x":0,"y":0,"numAPlusses":0},"Jamie Rup":{"id":"Jamie Rup","loc":0,"x":0,"y":0,"numAPlusses":0}},"maxPlayers":2,"difficulty":5,"countdownStarted":true,"countdown":0,"APlusLoc":40}]
+[{"roomCode":"AXVSTC","usernames":["guest_2190.10841996269"],"players":{"guest_2190.10841996269":{"id":"guest_2190.10841996269","loc":0,"x":0,"y":0,"numAPlusses":0}},"maxPlayers":2,"difficulty":5,"countdownStarted":false,"countdown":10,"APlusLoc":21}]
